Add warnings about unsupported B-rank/ARR hunts
The `/phb next` command now prints a (suppressable) warning to your chat when an elite mark is chosen, informing you that such targets do not have locations in the plugin. The main window also warns of this, as well as the fact that hunts from A Realm Reborn are not supported.

Hopefully this will make users stop sending feedback asking for those two things to be added.

Also, my friend recently had a good five minutes of angry confusion about why their map flag wasn't updating when HB picked a new mark for them, and it wasn't until I thought to ask "does it say it picked an ELITE mark?" that the problem was found. Shoutout to you, buddy. You've been anonymously immortalised in a git commit message.